+++ testing/testing_parse_opts.cpp
@@ -11,22 +11,22 @@ int main( int argc, char** argv )
     opts.parse_opts( argc, argv );
     
     for( int itest = 0; itest < opts.ntest; ++itest ) {
-        printf( "m %5d, n %5d, k %5d\n",
-                (int) opts.msize[itest], (int) opts.nsize[itest], (int) opts.ksize[itest] );
+        printf( "m %5ld, n %5ld, k %5ld\n",
+                long(opts.msize[itest]), long(opts.nsize[itest]), long(opts.ksize[itest]) );
     }
     printf( "\n" );
     
-    printf( "ntest    %d\n", (int) opts.ntest );
+    printf( "ntest    %ld\n", long(opts.ntest) );
     printf( "\n" );
     
-    printf( "nb       %d\n", (int) opts.nb       );
-    printf( "nrhs     %d\n", (int) opts.nrhs     );
-    printf( "nqueue   %d\n", (int) opts.nqueue   );
-    printf( "ngpu     %d\n", (int) opts.ngpu     );
-    printf( "niter    %d\n", (int) opts.niter    );
-    printf( "nthread  %d\n", (int) opts.nthread  );
-    printf( "itype    %d\n", (int) opts.itype    );
-    printf( "verbose  %d\n", (int) opts.verbose  );
+    printf( "nb       %ld\n", long(opts.nb)       );
+    printf( "nrhs     %ld\n", long(opts.nrhs)     );
+    printf( "nqueue   %ld\n", long(opts.nqueue)   );
+    printf( "ngpu     %ld\n", long(opts.ngpu)     );
+    printf( "niter    %ld\n", long(opts.niter)    );
+    printf( "nthread  %ld\n", long(opts.nthread)  );
+    printf( "itype    %ld\n", long(opts.itype)    );
+    printf( "verbose  %ld\n", long(opts.verbose)  );
     printf( "\n" );
     
     printf( "check    %s\n", (opts.check  ? "true" : "false") );
@@ -44,7 +44,7 @@ int main( int argc, char** argv )
     printf( "jobvl    %3d (%s)\n", opts.jobvl,  lapack_vec_const(   opts.jobvl  ));
     
     for( auto iter = opts.svd_work.begin(); iter < opts.svd_work.end(); ++iter ) {
-        printf( "svd_work %d\n", (int) *iter );
+        printf( "svd_work %ld\n", long(*iter) );
     }
     for( auto iter = opts.jobu.begin(); iter < opts.jobu.end(); ++iter ) {
         printf( "jobu     %3d (%s)\n", *iter, lapack_vec_const( *iter ));
