+++ testing/testing_ztrmm.cpp
@@ -149,16 +149,16 @@ int main( int argc, char** argv)
                 blasf77_zaxpy( &sizeB, &c_neg_one, h_B, &ione, h_Bcublas, &ione );
                 cublas_error = lapackf77_zlange( "M", &M, &N, h_Bcublas, &ldb, work ) / Cnorm;
                 
-                printf("%5d %5d   %7.2f (%7.2f)   %7.2f (%7.2f)    %8.2e   %s\n",
-                       (int) M, (int) N,
+                printf("%5ld %5ld   %7.2f (%7.2f)   %7.2f (%7.2f)    %8.2e   %s\n",
+                       long(M), long(N),
                        cublas_perf, 1000.*cublas_time,
                        cpu_perf,    1000.*cpu_time,
                        cublas_error, (cublas_error < tol ? "ok" : "failed"));
                 status += ! (cublas_error < tol);
             }
             else {
-                printf("%5d %5d   %7.2f (%7.2f)    ---   (  ---  )    ---     ---\n",
-                       (int) M, (int) N,
+                printf("%5ld %5ld   %7.2f (%7.2f)    ---   (  ---  )    ---     ---\n",
+                       long(M), long(N),
                        cublas_perf, 1000.*cublas_time);
             }
